Create web component util
